Widen integer properties to 64 bits so unsigned 32-bit values survive
A 32-bit unsigned property does not fit in an int, so 0xFFFFFFFF came
back as -1: the right bits carried as the wrong number, which is this
format's characteristic failure. SendPropDecoder.ReadInt and
PropertyValue now use long, matching the reference parser's choice for
the same reason.

Confirmed against the oracle rather than by assertion. Properties 618 and
619 of z1800's snapshot 0 entity 3 now both report 4294967295 where this
parser said -1, and the per-snapshot value differential over the first
twenty snapshots is otherwise clean - 1,183 lines compared, six differing
and all six only in float rendering, where Rust prints -37.789063 and
C#'s round-trip format prints -37.789062 for the same float.

Sign extension deliberately still happens at int width before widening,
so a signed 32-bit -1 stays -1 rather than becoming 4294967295. That
control has its own test, because widening a return type is exactly the
change that would silently break it.

481 tests, 0 warnings.

+++ managed/Tf2DemoSalvage.Core/Schema/PropertyValue.cs
@@ -47,7 +47,7 @@ public enum PropertyValueKind : byte
 /// </remarks>
 public readonly record struct PropertyValue
 {
-    private readonly int _int;
+    private readonly long _int;
     private readonly float _x;
     private readonly float _y;
     private readonly float _z;
@@ -56,7 +56,7 @@ public readonly record struct PropertyValue
 
     private PropertyValue(
         PropertyValueKind kind,
-        int intValue = 0,
+        long intValue = 0,
         float x = 0f,
         float y = 0f,
         float z = 0f,
@@ -82,7 +82,11 @@ public readonly record struct PropertyValue
     /// <summary>Creates an integer value.</summary>
     /// <param name="value">The integer.</param>
     /// <returns>The wrapped value.</returns>
-    public static PropertyValue FromInt(int value) =>
+    /// <remarks>
+    /// 64-bit because a 32-bit unsigned property does not fit in an <see cref="int"/>. Narrower
+    /// storage reports 0xFFFFFFFF as -1: the right bits carried as the wrong number.
+    /// </remarks>
+    public static PropertyValue FromInt(long value) =>
         new(PropertyValueKind.Int, intValue: value);
 
     /// <summary>Creates a float value.</summary>
@@ -120,7 +124,7 @@ public readonly record struct PropertyValue
 
     /// <summary>The integer payload.</summary>
     /// <exception cref="InvalidOperationException">This value is not an integer.</exception>
-    public int AsInt => Kind == PropertyValueKind.Int
+    public long AsInt => Kind == PropertyValueKind.Int
         ? _int
         : throw Mismatch(PropertyValueKind.Int);
